fix lint

+++ grug/math/src/fixed_point.rs
@@ -19,22 +19,22 @@ pub trait FixedPoint<U> {
 // ------------------------------------ dec ------------------------------------
 
 impl<const S: u32> FixedPoint<u128> for Dec<u128, S> {
-    const PRECISION: Int<u128> = Int::<u128>::new(10u128.pow(S as u32));
+    const PRECISION: Int<u128> = Int::<u128>::new(10u128.pow(S));
     const TICK: Self = Self::raw(Int::<u128>::ONE);
 }
 
 impl<const S: u32> FixedPoint<i128> for Dec<i128, S> {
-    const PRECISION: Int<i128> = Int::<i128>::new(10i128.pow(S as u32));
+    const PRECISION: Int<i128> = Int::<i128>::new(10i128.pow(S));
     const TICK: Self = Self::raw(Int::<i128>::ONE);
 }
 
 impl<const S: u32> FixedPoint<U256> for Dec<U256, S> {
-    const PRECISION: Int<U256> = Int::<U256>::new_from_u128(10u128.pow(S as u32));
+    const PRECISION: Int<U256> = Int::<U256>::new_from_u128(10u128.pow(S));
     const TICK: Self = Self::raw(Int::<U256>::ONE);
 }
 
 impl<const S: u32> FixedPoint<I256> for Dec<I256, S> {
-    const PRECISION: Int<I256> = Int::<I256>::new_from_i128(10i128.pow(S as u32));
+    const PRECISION: Int<I256> = Int::<I256>::new_from_i128(10i128.pow(S));
     const TICK: Self = Self::raw(Int::<I256>::ONE);
 }
